+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # 1.x branch
 ## 1.0 branch
+### 1.0.0-beta2
+* added `ConstantEntity::getValueAsString()` method, fixed bug for class costants
+    with an array as value.
+
 ### 1.0.0-beta1
 * first release.